feat: rename task folder to activity
